Move scheduler queries into the new module
